requirements.txt
streamlit>=1.2.0
reportlab>=3.6.8
Pygments>=2.10.0
html2text>=2020.1.16
tqdm>=4.62.3
python-dotenv>=0.19.0
python-pptx>=0.6.21
pywin32>=305
Pillow>=8.4.0

